# server/docs/openapi.yaml
openapi: "3.1.0"
info:
  title: White-Label School Management System
  version: "3.4.0"
  description: >
    Multi-tenant school management API. Supports SuperAdmin (platform-level)
    and tenant-scoped Admin/Teacher/Student roles.
    v3.4.0: Student role added; attendance correction (PUT /attendance/{recordId});
    link-student-account (PUT /students/{studentId}/link-account);
    reactivate tenant (PUT /super-admin/tenants/{tenantId}/reactivate);
    createTenant now requires an admin block.
servers:
  - url: http://localhost:3000/api
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, timestamp]
          properties:
            code:
              type: string
              example: INVALID_CREDENTIALS
            message:
              type: string
              example: Invalid credentials
            details:
              type: object
              additionalProperties: true
            timestamp:
              type: string
              format: date-time
...
    Student:
      type: object
      required: [id, name, classId, batchId]
      properties:
        id:
          type: string
        name:
          type: string
        userId:
          type: string
          nullable: true
          description: Linked user account ID (v3.4 CR-08); null if not yet linked
        classId:
          type: string
        className:
          type: string
        batchId:
          type: string
        batchName:
          type: string
...
    BulkDeleteRequest:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 100
          uniqueItems: true
    BulkDeleteResponse:
      type: object
      required: [deleted, failed]
      properties:
        deleted:
          type: array
          items:
            type: string
        failed:
          type: array
          items:
            type: object
            required: [id, reason, message]
            properties:
              id:
                type: string
              reason:
                type: string
                enum: [NOT_FOUND, HAS_REFERENCES]
              message:
                type: string
    Pagination:
      type: object
      required: [limit, offset, total]
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: UNAUTHORIZED
              message: Missing or invalid token
              timestamp: "2026-02-26T01:00:00Z"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: FORBIDDEN
              message: Insufficient permissions
              timestamp: "2026-02-26T01:00:00Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Validation failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /students:
    get:
      summary: List students
      operationId: listStudents
      tags: [Students]
      parameters:
        - name: classId
          in: query
          schema:
            type: string
        - name: batchId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: Student list
          content:
            application/json:
              schema:
                type: object
                required: [students, pagination]
                properties:
                  students:
                    type: array
                    items:
                      $ref: "#/components/schemas/Student"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      summary: Create student
      operationId: createStudent
      tags: [Students]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, classId, batchId]
              properties:
                name:
                  type: string
                  maxLength: 255
                classId:
                  type: string
                batchId:
                  type: string
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                type: object
                required: [student]
                properties:
                  student:
                    $ref: "#/components/schemas/Student"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /students/bulk:
    delete:
      summary: Bulk soft-delete students
      operationId: bulkDeleteStudents
      tags: [Students]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkDeleteRequest"
      responses:
        "200":
          description: Bulk delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkDeleteResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /students/{id}:
    delete:
      summary: Soft-delete a student
      operationId: deleteStudent
      tags: [Students]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  # v3.4 CR-08: Link a student record to a user account
  /students/{studentId}/link-account:
    put:
      summary: Link student record to a user account
      operationId: linkStudentAccount
      tags: [Students]
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId:
                  type: string
                  description: ID of the User to link to this student record
            examples:
              success:
                value:
                  userId: U-abc123
      responses:
        "200":
          description: Student linked to user account
          content:
            application/json:
              schema:
                type: object
                required: [student]
                properties:
                  student:
                    $ref: "#/components/schemas/Student"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Student already linked or user already linked to another student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                already_linked:
                  summary: Student already has a linked account
                  value:
                    error:
                      code: USER_ALREADY_LINKED
                      message: This student already has a linked user account
                      timestamp: "2026-02-26T07:00:00Z"
...